+++ CHANGELOG.rst
@@ -13,6 +13,7 @@ Changelog
 * Update :command:`import-read-depth` command to accept archive file as input instead of TSV file.
 * Add ``fitted`` argument to ``api.plot.plot_bam_copy_number`` method.
 * From now on, missing copy number will be imputed with forward filling instead of column median.
+* Update :command:`predict-cnv` command to support a user-defined CNV caller.
 
 0.4.1 (2021-09-21)
 ------------------